moving uploadInterfaceDocument() from common.ts to data-preparation-service.ts
